+++ src/Dap.Remote.Web/App/Helper.fs
@@ -8,10 +8,10 @@ open Microsoft.Extensions.DependencyInjection
 open Dap.Prelude
 open Dap.Platform
 
-let runWebApp<'app when 'app :> IBaseApp> (newWebHost : 'app -> WebHost) (app : 'app) =
+let runWebApp<'app when 'app :> IBaseApp> (newWebHost : 'app -> WebHost) (cleanup : unit -> unit) (app : 'app) =
     let webHost = newWebHost app
     let host =
         new WebHostBuilder ()
         |> WebHostBuilderKestrelExtensions.UseKestrel
         |> webHost.Build
-    WebApp.WebApp<'app>.Run host app
+    WebApp.WebApp<'app>.Run host cleanup app